Modified 5 file(s): src/fn.cpp, src/global.cpp, src/ref.h, ...and 2 more

+++ src/fn.cpp
@@ -12,7 +12,7 @@ struct FnImpl {
 };
 
 Fn::Fn() {
-	p = new FnImpl(voidTy(), 0, {}, {});
+	p = new FnImpl(voidTy(), (size_t)0, {}, {});
 }
 
 Fn::Fn(Type rty, const Ref& ref, const vector<Term>& params) {
